# compression: fix bad_allocs in lz4 compression, gzip compression+decompression