--- a/src/v/compression/internal/lz4_frame_compressor.cc
+++ b/src/v/compression/internal/lz4_frame_compressor.cc
@@ -10,11 +10,9 @@
 #include "compression/internal/lz4_frame_compressor.h"
 
 #include "bytes/bytes.h"
-#include "compression/logger.h"
 #include "static_deleter_fn.h"
 #include "units.h"
 #include "vassert.h"
-#include "vlog.h"
 
 #include <seastar/core/temporary_buffer.hh>
 
@@ -76,39 +74,90 @@ iobuf lz4_frame_compressor::compress(const iobuf& b) {
     prefs.compressionLevel = 1; // default
     prefs.frameInfo = {
       .blockMode = LZ4F_blockIndependent, .contentSize = b.size_bytes()};
-    const size_t output_buffer_size = LZ4F_compressBound(b.size_bytes(), &prefs)
-                                      + lz4f_footer_size + lz4f_header_size;
-    check_lz4_error("lz4_compressbound erorr:{}", output_buffer_size);
-    ss::temporary_buffer<char> obuf(output_buffer_size);
-    char* out = obuf.get_write();
-    LZ4F_errorCode_t code = LZ4F_compressBegin(
-      ctx, out, output_buffer_size, &prefs);
-    check_lz4_error("lz4f_compressbegin error:{}", code);
 
-    // start after the bytes from compressBegin
-    size_t consumed_bytes = code;
-    for (auto& frag : b) {
+    const size_t max_chunk_size = details::io_allocation_size::max_chunk_size;
+
+    const size_t compress_bound = LZ4F_compressBound(b.size_bytes(), &prefs);
+    check_lz4_error("lz4_compressbound error:{}", compress_bound);
+
+    const size_t estimated_output_size = compress_bound + lz4f_footer_size
+                                         + lz4f_header_size;
+
+    size_t output_chunk_size = std::min(max_chunk_size, estimated_output_size);
+
+    ss::temporary_buffer<char> obuf(output_chunk_size);
+    char* output = obuf.get_write();
+    size_t output_sz = output_chunk_size;
+    LZ4F_errorCode_t code = LZ4F_compressBegin(
+      ctx, output, obuf.size(), &prefs);
+    check_lz4_error("lz4f_compressbegin error:{}", code);
+    size_t output_cursor = code;
+
+    auto frag_i = b.begin();
+
+    const char* input = nullptr;
+    size_t input_sz = 0;
+    size_t input_cursor = 0;
+    if (frag_i != b.end()) {
+        input = frag_i->get();
+        input_sz = frag_i->size();
+    }
+
+    // We do not consume entire input chunks at once, to avoid
+    // max_chunk_size input chunks resulting in >max_chunk_size output
+    // chunks.  A half-sized input chunk never results in a LZ4F_compressBound
+    // that exceeds a the max output chunk.
+    const size_t max_input_chunk_size = max_chunk_size / 2;
+
+    iobuf ret;
+    while (input_cursor < input_sz || frag_i != b.end()) {
+        if ((input_sz - input_cursor) == 0 && frag_i != b.end()) {
+            frag_i++;
+            if (frag_i != b.end()) {
+                input = frag_i->get();
+                input_sz = frag_i->size();
+                input_cursor = 0;
+            }
+        }
+
+        size_t input_chunk_size = std::min(
+          input_sz - input_cursor, max_input_chunk_size);
+
         code = LZ4F_compressUpdate(
           ctx,
-          // NOLINTNEXTLINE
-          out + consumed_bytes,
-          output_buffer_size - consumed_bytes,
-          frag.get(),
-          frag.size(),
+          output + output_cursor,
+          output_sz - output_cursor,
+          input + input_cursor,
+          input_chunk_size,
           nullptr);
         check_lz4_error("lz4f_compressupdate error:{}", code);
-        consumed_bytes += code;
+
+        // We always consume all input buffer
+        input_cursor += input_chunk_size;
+
+        // Advance by how many bytes we consumed
+        output_cursor += code;
+
+        input_chunk_size = std::min(
+          input_sz - input_cursor, max_input_chunk_size);
+        size_t next_output_size = LZ4F_compressBound(input_chunk_size, &prefs);
+
+        if (output_sz - output_cursor < next_output_size) {
+            obuf.trim(output_cursor);
+            ret.append(std::move(obuf));
+            output_chunk_size = std::min(max_chunk_size, output_chunk_size * 2);
+            obuf = ss::temporary_buffer<char>(output_chunk_size);
+            output = obuf.get_write();
+            output_sz = obuf.size();
+            output_cursor = 0;
+        }
     }
+
     code = LZ4F_compressEnd(
-      ctx,
-      // NOLINTNEXTLINE
-      out + consumed_bytes,
-      output_buffer_size - consumed_bytes,
-      nullptr);
+      ctx, output + output_cursor, output_sz - output_cursor, nullptr);
     check_lz4_error("lz4f_compressend:{}", code);
-    consumed_bytes += code;
-    obuf.trim(consumed_bytes);
-    iobuf ret;
+    output_cursor += code;
+    obuf.trim(output_cursor);
     ret.append(std::move(obuf));
     return ret;
 }
@@ -121,66 +170,78 @@ compute_frame_uncompressed_size(size_t frame_size, size_t original) {
     return frame_size;
 }
 
-static iobuf do_uncompressed(const char* src, const size_t src_size) {
-    auto ctx_ptr = make_decompression_context();
-    LZ4F_decompressionContext_t ctx = ctx_ptr.get();
-    LZ4F_frameInfo_t fi;
-    size_t in_sz = src_size;
-    LZ4F_errorCode_t code = LZ4F_getFrameInfo(ctx, &fi, src, &in_sz);
-    check_lz4_error("lz4f_getframeinfo error: {}", code);
-    size_t this_chunk_size = compute_frame_uncompressed_size(
-      fi.contentSize, src_size);
+iobuf lz4_frame_compressor::uncompress(iobuf const& input) {
+    size_t src_size = input.size_bytes();
 
-    size_t max_chunk = 128_KiB;
-    this_chunk_size = std::min(this_chunk_size, max_chunk);
-
-    // We will decompress into temporary_buffers, and each time
-    // we spill past estimated_output_size, push the
-    ss::temporary_buffer<char> obuf(this_chunk_size);
-    char* out = obuf.get_write();
-
-    // Offset in `src` buffer
-    size_t src_offset = in_sz;
-
-    // Offset in `out` buffer
-    size_t output_this_chunk{0};
+    const size_t max_chunk = details::io_allocation_size::max_chunk_size;
 
     // Compose fragmented result from a series of `obuf` temporary buffers
     iobuf ret;
 
-    while (src_offset < src_size) {
+    // Read interators/offsets
+    auto frag_i = input.begin();
+    size_t read_this_chunk{0};
+    size_t read_total{0};
+
+    auto ctx_ptr = make_decompression_context();
+    LZ4F_decompressionContext_t ctx = ctx_ptr.get();
+
+    // Prior to main loop, optionally consume header to learn total size
+    LZ4F_errorCode_t code = 0;
+    size_t decompressed_size = 0;
+    if (!input.empty() && input.begin()->size() >= lz4f_header_size) {
+        // This is optional, because it is not guaranteed that the caller
+        // passes us an iobuf with the header in a contiguous chunk.
+        auto& first_chunk = *(input.begin());
+        size_t sz_scratch = first_chunk.size();
+        LZ4F_frameInfo_t fi;
+        code = LZ4F_getFrameInfo(ctx, &fi, first_chunk.get(), &sz_scratch);
+        read_this_chunk = sz_scratch;
+        read_total += sz_scratch;
+        check_lz4_error("lz4f_getframeinfo error: {}", code);
+        decompressed_size = fi.contentSize;
+    }
+    size_t write_chunk_size = compute_frame_uncompressed_size(
+      decompressed_size, src_size);
+    write_chunk_size = std::min(write_chunk_size, max_chunk);
+
+    // Write buffer/offsets
+    size_t write_this_chunk{0};
+    ss::temporary_buffer<char> obuf(write_chunk_size);
+    char* out = obuf.get_write();
+
+    while (frag_i != input.end()) {
         // These variables at as an input to LZ4F_decompress to specify
         // buffer size, and then as an output for how many bytes the
         // buffers advanced.
-        size_t consumed_this_iteration = src_size - src_offset;
-        size_t output_this_iteration = this_chunk_size - output_this_chunk;
+        size_t consumed_this_iteration = frag_i->size() - read_this_chunk;
+        size_t output_this_iteration = obuf.size() - write_this_chunk;
 
         code = LZ4F_decompress(
           ctx,
-          // NOLINTNEXTLINE
-          out + output_this_chunk,
+          out + write_this_chunk,
           &output_this_iteration,
-          // NOLINTNEXTLINE
-          src + src_offset,
+          frag_i->get() + read_this_chunk,
           &consumed_this_iteration,
           nullptr);
 
-        output_this_chunk += output_this_iteration;
-        src_offset += consumed_this_iteration;
+        write_this_chunk += output_this_iteration;
+        read_this_chunk += consumed_this_iteration;
+        read_total += consumed_this_iteration;
 
         vassert(
-          output_this_chunk <= this_chunk_size,
+          write_this_chunk <= write_chunk_size,
           "Appended more bytes that allowed. Max:{}, consumed:{} ({} this "
           "iteration)",
-          this_chunk_size,
-          output_this_chunk,
+          write_chunk_size,
+          write_this_chunk,
           output_this_iteration);
 
         vassert(
-          src_offset <= src_size,
+          read_this_chunk <= frag_i->size(),
           "Consumed more bytes than in input buffer (size {}, consumed {})",
-          src_size,
-          src_offset);
+          frag_i->size(),
+          read_this_chunk);
 
         check_lz4_error("lz4f_decompress error: {}", code);
 
@@ -188,45 +249,37 @@ static iobuf do_uncompressed(const char* src, const size_t src_size) {
             break;
         }
 
-        if (output_this_chunk == this_chunk_size && src_offset < src_size) {
+        if (read_this_chunk == frag_i->size()) {
+            read_this_chunk = 0;
+            frag_i++;
+        }
+
+        if (write_this_chunk == write_chunk_size && frag_i != input.end()) {
             ret.append(std::move(obuf));
 
             // If we were using a smaller-than-max chunk, grow it
-            this_chunk_size = std::min(max_chunk, this_chunk_size * 2);
+            write_chunk_size = std::min(max_chunk, write_chunk_size * 2);
 
-            obuf = ss::temporary_buffer<char>(this_chunk_size);
+            obuf = ss::temporary_buffer<char>(write_chunk_size);
             out = obuf.get_write();
-            output_this_chunk = 0;
+            write_this_chunk = 0;
         }
     }
 
-    if (unlikely(src_offset < src_size)) {
+    if (unlikely(read_total < src_size)) {
         throw std::runtime_error(fmt::format(
           "lz4 error. could not consume all input bytes in decompression. "
           "Input:{}, consumed:{}",
           src_size,
-          src_offset));
+          read_total));
     }
 
-    if (output_this_chunk > 0) {
-        obuf.trim(output_this_chunk);
+    if (write_this_chunk > 0) {
+        obuf.trim(write_this_chunk);
         ret.append(std::move(obuf));
     }
 
     return ret;
-}
-
-iobuf lz4_frame_compressor::uncompress(const iobuf& b) {
-    if (std::distance(b.begin(), b.end()) == 1) {
-        return do_uncompressed(b.begin()->get(), b.size_bytes());
-    }
-    // linearize buffer
-    // TODO: optimize iobuf
-    auto linearized = iobuf_to_bytes(b);
-    return do_uncompressed(
-      // NOLINTNEXTLINE
-      reinterpret_cast<const char*>(linearized.data()),
-      linearized.size());
 }
 
 } // namespace compression::internal