Previous PR https://github.com/redpanda-data/redpanda/pull/9563 fixed the most impactful case, but there were several other places in compression code where we could crash on oversized allocations.  This PR revises compression and decompression paths to keep their allocations inside 128KiB.

Performance:
- gzip decompression is >2x better, because the old implementation did two full decompressions (one to learn the size, one to write the data).
- LZ4 is noticeably slower: as a faster algorithm it makes the impact of doing more buffer management most apparent.  I tried turning `max_chunk_size` up to 1MiB and the performance of the 1MiB microbenchmark goes back to exactly what it was pre-fixes, so the performance change appears to be entirely attributable to the (necessary) fragmentation, rather than being anything pathological about the new code.

```

# Before fixes

lz4_1mb.compress                               15024    50.629us     3.845ns    50.625us    50.667us       5.000       0.000   1125423.4
lz4_10mb.compress                               1600   481.005us    36.839ns   480.945us   481.280us       5.000       0.000  11213003.1
lz4_1mb.uncompress                             10626    27.361us    16.434ns    27.342us    27.419us       5.000       0.000    822936.3
lz4_10mb.uncompress                             1123   268.318us    31.929ns   268.163us   268.955us       5.000       0.000   8211884.1
gzip_1mb.compress                                600     1.659ms     1.850us     1.655ms     1.661ms       7.000       0.000  37183003.3
gzip_10mb.compress                                61    16.493ms    18.139us    16.371ms    16.511ms       7.000       0.000 371541188.2
gzip_1mb.uncompress                              389   883.695us     1.419us   882.276us   886.345us       5.000       0.000  15696291.6
gzip_10mb.uncompress                              40     8.708ms     3.876us     8.704ms     8.722ms       5.000       0.000 156360045.5

# After fixes

lz4_1mb.compress                               13712    56.909us    47.060ns    56.862us    57.121us       5.000       0.000   1130008.4
lz4_10mb.compress                               1418   561.559us   239.045ns   559.300us   562.755us       8.000       0.000  11247098.7
lz4_1mb.uncompress                              8892    40.528us    41.428ns    40.479us    40.612us      26.000       0.000    830291.6
lz4_10mb.uncompress                              882   433.802us    49.307ns   433.599us   434.463us     242.000       0.000   8295986.0

gzip_1mb.compress                                595     1.665ms     1.768us     1.663ms     1.667ms       7.000       0.000  37183299.4
gzip_10mb.compress                                61    16.548ms    40.417us    16.467ms    16.630ms       7.000       0.000 371542070.3
gzip_1mb.uncompress                              491   375.450us    84.352ns   375.198us   376.265us      28.000       0.000   7682711.6
gzip_10mb.uncompress                              50     3.773ms    10.248us     3.761ms     3.784ms     241.000       0.000  76618644.5

```

## Backports Required

- [ ] none - not a bug fix
- [ ] none - this is a backport
- [ ] none - issue does not exist in previous branches
- [ ] none - papercut/not impactful enough to backport
- [x] v23.1.x
- [x] v22.3.x
- [x] v22.2.x

(And for v22.3.x + v22.2.x, cherry pick in https://github.com/redpanda-data/redpanda/pull/9563 as well)

## Release Notes

### Bug Fixes

* Fix cases where sending very large batches to compacted topics using LZ4 or gzip compression could result in bad_alloc errors.