@dotnwat 2023-03-23
>gzip decompression is >2x better, because the old implementation did two full decompressions (one to learn the size, one to write the data).

omg