@jcsp 2023-03-29
- https://github.com/redpanda-data/redpanda/issues/8496
- https://github.com/redpanda-data/redpanda/issues/7756