@jcsp 2023-03-29
/backport v23.1.x