@jcsp 2023-03-29
/backport v22.2.x