@vbotbuildovich 2023-03-29
Failed to run cherry-pick command. I executed the below command:

```
git cherry-pick -x 423199e1c14c8d6c5590abedfda44f99aa9a97cd 14562d7250adcff3e3213823fdedf550e2cfc7f1 b847f0c202abf1259ae147aad7d25c014f1bd258 01b9962dc25a194909e69a98d6b46d5356dbb78b 3a2ee8349ede369c32d0947ff9ed89e551b7de6c
```

[Workflow run logs](https://github.com/redpanda-data/redpanda/actions/runs/4557243510).